Ripped out /setting POST and GET
These end points (imo) should be defined in the main application instead of being hardcoded in the lib.
If the user doesn't need them then they are left over endpoints in the lib, If a user wants to expand them then the hard coded ones are left in the lib.