Ripped out /Scan GET
This should be handled by the main application. This built in /scan endpoint is broken. It used a fixed sized json doc which overflowed if used in an environment with more than a handle of networks.
If this is used in an low populated area then this is fine, but these days even ISP issued routers are coming with WiFi extenders which increases the AP's in an avg home from 1 to 3.
This means if this lib is used where the device can see just one neighbour with a lengthy SSID and a few extenders the build in /scan endpoint will return an incomplete and or broken list of AP's

Moving this to main will allow for easier changes to the /scan endpoint.