Added an begin() which dopesn't need a config.
the app side config is not always used by applications. So adding an option to save having to put in a empty config into the begin just to get the config manager working.
this shouldn't be a breaking change as the other read/write config calls check the config size. I'll have to look at the rest of the code to make sure, but atm it will do for what I need it to do.